refactor(ledger): isolate failure fence policy
